requirements.txt
newspaper3k==0.1.9
watson-developer-cloud==0.23.0
requests==2.13.0
Flask==0.10.1
gunicorn==19.6.0
flask-socketio==2.8.4
eventlet==0.20.1
markovify==0.5.4